+++ src/onegov.reservation/onegov/reservation/tests/test_pricing.py
@@ -17,8 +17,10 @@ def test_per_hour_pricing(libres_context):
     token = scheduler.reserve('info@example.org', dates)
     reservation = scheduler.reservations_by_token(token).one()
 
-    assert reservation.price.amount == 60
-    assert reservation.price.currency == 'CHF'
+    assert reservation.price().amount == 60
+    assert reservation.price().currency == 'CHF'
+    assert reservation.price(resource).amount == 60
+    assert reservation.price(resource).currency == 'CHF'
 
 
 def test_multiple_allocations(libres_context):
@@ -39,10 +41,14 @@ def test_multiple_allocations(libres_context):
     token = scheduler.reserve('info@example.org', dates)
     reservations = scheduler.reservations_by_token(token).all()
 
-    assert reservations[0].price.amount == 60
-    assert reservations[0].price.currency == 'CHF'
-    assert reservations[1].price.amount == 60
-    assert reservations[1].price.currency == 'CHF'
+    assert reservations[0].price().amount == 60
+    assert reservations[0].price().currency == 'CHF'
+    assert reservations[1].price().amount == 60
+    assert reservations[1].price().currency == 'CHF'
+    assert reservations[0].price(resource).amount == 60
+    assert reservations[0].price(resource).currency == 'CHF'
+    assert reservations[1].price(resource).amount == 60
+    assert reservations[1].price(resource).currency == 'CHF'
 
 
 def test_per_reservation_pricing(libres_context):
@@ -60,5 +66,7 @@ def test_per_reservation_pricing(libres_context):
     token = scheduler.reserve('info@example.org', dates, quota=2)
     reservation = scheduler.reservations_by_token(token).one()
 
-    assert reservation.price.amount == 20
-    assert reservation.price.currency == 'CHF'
+    assert reservation.price().amount == 20
+    assert reservation.price().currency == 'CHF'
+    assert reservation.price(resource).amount == 20
+    assert reservation.price(resource).currency == 'CHF'